Refactor: HabtiPayload -> HabitPayload로 이름 수정
